moved LERCCodec to lib so that I can use unary operators
