+++ web/js/lib/LERCCodec.js
